requirements.txt
streamlit==1.36.0
reportlab==4.2.2
Pillow==10.4.0
numpy==2.0.1
opencv-python-headless==4.10.0.84
stripe==10.12.0
